requirements.txt
streamlit
pytesseract
pdf2image
Pillow
pydantic
requests
python-multipart
numpy
tqdm
